+++ lib/features/auth/presentation/auth_gate.dart
@@ -2,12 +2,17 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_smart_dialog/flutter_smart_dialog.dart'
     show SmartDialog;
+import 'package:lucide_icons_flutter/lucide_icons.dart';
 import 'package:pudding/common/components/app_info_cont.dart';
+import 'package:pudding/common/components/btns.dart'
+    show defaultTextIconBtn, filledTextIconBtn;
 import 'package:pudding/common/components/full_page_loading.dart';
 import 'package:pudding/common/components/view_wrappers.dart';
+import 'package:pudding/common/parts.dart';
 import 'package:pudding/core/models/appbar_cfg_model.dart';
 import 'package:pudding/features/auth/presentation/signin_anony.dart';
 import 'package:pudding/features/auth/presentation/signin_email.dart';
+import 'package:pudding/features/auth/presentation/signup_email.dart';
 import 'package:pudding/features/auth/providers/auth_providers.dart';
 import 'package:pudding/features/collections/presentation/collections_list_screen.dart';
 
@@ -22,13 +27,14 @@ class AuthGate extends ConsumerWidget {
         if (user == null) {
           return pageViewWrapper(
             extendBodyBehindAppBar: true,
-            extendBody: true,
+            extendBody: false,
             showEndDrawer: false,
             appBarCfg: AppbarCfgModel(
               titleStr: "Welcome!",
               overrideActions: const [],
             ),
             body: loginOptionsSelectionView(),
+            floatingActionButton: renderSignupEmailBtn(),
             bottomSheet: AppInfoCont(),
           );
         } else {
@@ -59,29 +65,46 @@ class AuthGate extends ConsumerWidget {
     );
   }
 
+  Padding renderSignupEmailBtn() {
+    return Padding(
+      padding: Parts.bigEdgeInsetsAll,
+      child: filledTextIconBtn(
+        // TODO: localize
+        text: const Text("Create account"),
+        icon: const Icon(LucideIcons.plus),
+        onPressed: () async => await SmartDialog.show(
+          builder: (_) => const SignupEmailPanel(),
+          alignment: Alignment.topCenter,
+        ),
+      ),
+    );
+  }
+
   ElevatedButton renderSigninAnonyBtn() {
-    return ElevatedButton.icon(
-      onPressed: () => SmartDialog.show(
+    return defaultTextIconBtn(
+      onPressed: () async => await SmartDialog.show(
         keepSingle: true,
-        builder: (context) => SigninAnonyPanel(),
+        builder: (_) => const SigninAnonyPanel(),
         alignment: Alignment.centerLeft,
         usePenetrate: true,
         clickMaskDismiss: false,
-      ),
-      label: const Text("Sign in anonymously"),
+      ), //TODO: localize
+      text: const Text("Sign in anonymously"),
+      icon: const Icon(LucideIcons.hatGlasses),
     );
   }
 
   ElevatedButton renderSigninEmailBtn() {
-    return ElevatedButton.icon(
-      onPressed: () => SmartDialog.show(
+    return defaultTextIconBtn(
+      onPressed: () async => await SmartDialog.show(
         keepSingle: true,
-        builder: (context) => SigninEmailPanel(),
+        builder: (_) => const SigninEmailPanel(),
         alignment: Alignment.centerRight,
         usePenetrate: true,
         clickMaskDismiss: false,
-      ),
-      label: const Text("Sign in with email"),
+      ), //TODO: localize
+      text: const Text("Sign in with email"),
+      icon: const Icon(LucideIcons.mail),
     );
   }
 }